Move org rotation into a global plugin
