+++ components/PetitionForm.vue
@@ -126,7 +126,7 @@
         <span v-else>Write Congress</span>
       </button>
       <no-ssr>
-        <disclaimer :org="org"></disclaimer>
+        <disclaimer></disclaimer>
       </no-ssr>
     </form>
     <div class="petition-copy">
